mangsir 19

+++ backend/utils/removeImg.js
@@ -1,5 +1,13 @@
 import fs from "fs";
 export const removeImg = (path) => {
-    fs.unlinkSync(path);
+    try {
+        if (fs.existsSync(path)) {
+            fs.unlinkSync(path);
+        }
+    } catch (error) {
+        console.log(`file not found:${path}`)
+
+    }
+
 
 }
